1. some tc util refactoring for tc content management. 2. fixed the Util.parseFloat method and return a double precision value.

+++ JonDonym/trunk/src/anon/infoservice/ServiceLocation.java
@@ -132,7 +132,7 @@ public class ServiceLocation
 						try
 						{
 							longitude = coordinates.elementAt(0).toString();
-							Util.parseFloat(longitude);
+							Util.parseDouble(longitude);
 							longitude = longitude.trim();
 						}
 						catch (NumberFormatException a_e)
@@ -142,7 +142,7 @@ public class ServiceLocation
 						try
 						{
 							latitude = coordinates.elementAt(1).toString();
-							Util.parseFloat(latitude);
+							Util.parseDouble(latitude);
 							latitude = latitude.trim();
 						}
 						catch (NumberFormatException a_e)
